requirements.txt
setuptools
flask
google-cloud-bigquery
google-cloud